baselines/baseline_requirements.txt
numpy
pandas
opencv-python
networkx
h5py
tqdm
vocab
revtok
Pillow
tensorboardX
pyyaml
matplotlib
transformers
scikit-learn
torchmetrics
torchtext==0.12.0
graphviz
pydot
sentencepiece
sentence-transformers
open_clip_torch
git+https://github.com/openai/CLIP.git